api/requirements.txt
boto3==1.29.0
beautifulsoup4==4.12.2
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
chardet==5.2.0
markdown==3.5.1 